requirements.txt
imaplib2
pyzbar
pillow
google-api-python-client
google-auth
google-auth-oauthlib
